Deprecated the LiquidState public constructor
- I should have made this internal from the beginning but it was an oversight. Users have no reason to use this constructor and it would only cause issues if using it as a default param in Composables instead of `rememberLiquidState` as it won't survive recompositions.
- Also took this opportunity to move the Liquefiable state class into the internal directory since it is an internal class. This was at one time public which is why it was there, but fortunately it was made internal before any release.
